+++ biweeklybudget/vendored/install_vendored.sh
@@ -12,10 +12,4 @@ pip install --upgrade --target . --no-deps git+https://github.com/jantman/ofxcli
 curl -o ofxclient/LICENSE https://raw.githubusercontent.com/jantman/ofxclient/afa8d2175483bf4f50632179f434021782f49d9c/LICENSE
 curl -o ofxclient/setup.py.src https://github.com/jantman/ofxclient/raw/afa8d2175483bf4f50632179f434021782f49d9c/setup.py
 
-# to incorporate https://github.com/jseutter/ofxparse/pull/127 which was merged but not released
-# licensed under the MIT License
-pip install --upgrade --target . --no-deps git+https://github.com/jseutter/ofxparse.git@19c04b40e2c7c3cb2943344f6108bcaf3d968725#egg=ofxparse
-curl -o ofxparse/LICENSE https://raw.githubusercontent.com/jseutter/ofxparse/19c04b40e2c7c3cb2943344f6108bcaf3d968725/LICENSE
-curl -o ofxparse/setup.py.src https://raw.githubusercontent.com/jseutter/ofxparse/19c04b40e2c7c3cb2943344f6108bcaf3d968725/setup.py
-
 sed -i -r 's/^([[:space:]]*)from ofxclient/\1from biweeklybudget.vendored.ofxclient/g' ofxclient/*.py